split and rename files
